moved time finding heuristics from Calendar to EventScheduler
